requirements.txt
Flask
pymongo
gunicorn
elasticSearch
spacy
nltk
numpy
scikit-learn
pandas
sentence-transformers
transformers
requests
PyPDF2
pdf2image
pytesseract
Pillow
werkzeug
BeautifulSoup4
